bolt related stuff moved to separate package
